+++ src/main/java/com/optimus/dto/results/UserResultDTO.java
@@ -9,7 +9,7 @@ import java.util.Date;
 @Data
 public class  UserResultDTO implements Serializable {
     private String id;
-    private BigInteger userId;
+    private Long userId;
     private String userName;
     private String passWord;
     private String nickName;
